Hide event code in form if no environment value set and move useCountDown to same file as CountDown
